[FIX] account: don't unlink all unused attachment
Currently, when receiving an email with several attachments. Only XMLs
and PDFs are kept, the rest are unlink.

However, in some cases, we want to keep those attachment:
- When receiving a mail with 1 PDF and non-XMLs attachments
- When receiving a mail with 1 XML and non-PDFs attachments
- When receiving a mail with 1 PDF and 1 XML and other type-attachments

In those situtations, only 1 invoice will be created instead of several
(1/attachment). For those, we keep all the attachments and link them to
the unique invoice.

+++ addons/account/models/account_move.py
@@ -4762,8 +4762,11 @@ class AccountMove(models.Model):
             for invoice in invoices:
                 attachments_per_invoice[invoice] |= attachment
 
-        # Unlink the unused attachments
-        (attachments - attachments_in_invoices).unlink()
+        if len(attachments_per_invoice) == 1:
+            attachments_per_invoice = {self: attachments}
+        else:
+            # Unlink the unused attachments
+            (attachments - attachments_in_invoices).unlink()
 
         for invoice, attachments in attachments_per_invoice.items():
             if invoice == self:
